requirements.txt
marshmallow-mongoengine==0.9.1
mongoengine==0.16.2
pymongo==3.10.0
python-pcre==0.7
six==1.13.0
tipi-data==1.0.0
